# openapi.yaml
openapi: 3.0.3
info:
  title: Profile Service API
  description: Microservice for managing user profiles for COMP2001 CW2
  version: "1.0.0"

servers:
  - url: "http://localhost:8000"
    description: Local development

tags:
  - name: Health
  - name: Meta
  - name: Roles
  - name: Users
  - name: Profiles

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      responses:
        "200":
          description: Service is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok

  /openapi:
    get:
      tags: [Meta]
      summary: OpenAPI hint endpoint (points to openapi.yaml in the repository)
      parameters:
        - $ref: "#/components/parameters/XUser"
      responses:
        "200":
          description: Hint returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenApiHintResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /roles/me:
    get:
      tags: [Roles]
      summary: Get my username and role
      parameters:
        - $ref: "#/components/parameters/XUser"
      responses:
        "200":
          description: Role returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyRoleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /users:
    post:
      tags: [Users]
      summary: Create a new user profile (new user requirement)
      parameters:
        - $ref: "#/components/parameters/XUser"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileCreate"
      responses:
        "201":
          description: User profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileCreatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"

  /profiles:
    get:
      tags: [Profiles]
      summary: Get all profiles (admin/staff only)
      parameters:
        - $ref: "#/components/parameters/XUser"
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"

    post:
      tags: [Profiles]
      summary: Create a new profile
      parameters:
        - $ref: "#/components/parameters/XUser"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileCreate"
      responses:
        "201":
          description: Profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileCreatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"

  /profiles/{profile_id}:
    get:
      tags: [Profiles]
      summary: Get a profile by ID
      parameters:
        - $ref: "#/components/parameters/XUser"
        - name: profile_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

    put:
      tags: [Profiles]
      summary: Update a profile (partial update)
      parameters:
        - $ref: "#/components/parameters/XUser"
        - name: profile_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

    delete:
      tags: [Profiles]
      summary: Delete a profile by ID
      parameters:
        - $ref: "#/components/parameters/XUser"
        - name: profile_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Profile deleted (no content)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

components:
  parameters:
    XUser:
      name: X-User
      in: header
      required: true
      schema:
        type: string
      description: Username validated via the Authenticator API

  responses:
    Unauthorized:
      description: Missing or invalid X-User header / Authenticator validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Unauthorized user
    Forbidden:
      description: User does not have permission for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Profile not found
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Missing fields
    ServerError:
      description: Database/server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Database error

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
      required: [status]

    OpenApiHintResponse:
      type: object
      properties:
        message:
          type: string
        auth_header:
          type: string
        user:
          type: object
      required: [message, auth_header]

    MyRoleResponse:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
          enum: [admin, staff, user]
      required: [username, role]

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    Profile:
      type: object
      properties:
        ProfileID:
          type: integer
        Username:
          type: string
        Email:
          type: string
        Location:
          type: string
        PreferredActivity:
          type: string
        DateOfBirth:
          type: string
          format: date

    ProfileCreate:
      type: object
      required: [Username, Email, Location, PreferredActivity, DateOfBirth]
      properties:
        Username:
          type: string
        Email:
          type: string
          format: email
        Location:
          type: string
        PreferredActivity:
          type: string
        DateOfBirth:
          type: string
          format: date

    ProfileUpdate:
      type: object
      properties:
        Location:
          type: string
        PreferredActivity:
          type: string

    ProfileCreatedResponse:
      type: object
      properties:
        message:
          type: string
        ProfileID:
          type: integer
      required: [message, ProfileID]